% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add_header.R
\name{docx_add_header}
\alias{docx_add_header}
\title{add Header}
\usage{
docx_add_header(
  header = "This is Header",
  level = 1,
  size = 16,
  align = "left",
  color = "black",
  bold = TRUE,
  family = "Arial"
)
}
\arguments{
\item{header}{This is Header}

\item{level}{one numeric defaut is}

\item{size}{16}

\item{align}{left}

\item{color}{black}

\item{bold}{TRUE}

\item{family}{Arial}
}
\description{
add Header
}
