% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add_subtitle.R
\name{docx_add_subtitle}
\alias{docx_add_subtitle}
\title{Title}
\usage{
docx_add_subtitle(
  title = "This is sub Title",
  size = 18,
  family = "Arial",
  bold = TRUE,
  color = "black",
  italic = FALSE,
  underlined = FALSE,
  empty_line = TRUE,
  vertical.align = "subscript"
)
}
\arguments{
\item{title}{This is sub Title}

\item{size}{size, default is 18}

\item{family}{font family, default is Arial}

\item{bold}{TRUE}

\item{color}{black}

\item{italic}{FLASE}

\item{underlined}{FALSE}

\item{empty_line}{TRUE}

\item{vertical.align}{baseline}
}
\description{
Title
}
