% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add_text.R
\name{docx_add_text}
\alias{docx_add_text}
\title{add text}
\usage{
docx_add_text(
  text,
  size = 10,
  align = "left",
  color = "black",
  bold = FALSE,
  family = "Arial"
)
}
\arguments{
\item{text}{text}

\item{size}{12}

\item{align}{left}

\item{color}{black}

\item{bold}{FALSE}

\item{family}{Arial}
}
\description{
add text
}
