% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add_title.R
\name{docx_add_title}
\alias{docx_add_title}
\title{Title}
\usage{
docx_add_title(
  title = "This is Title",
  size = 15,
  family = "Arial",
  bold = TRUE,
  color = "black",
  italic = FALSE,
  underlined = FALSE,
  empty_line = TRUE,
  vertical.align = "baseline"
)
}
\arguments{
\item{title}{This is Title}

\item{size}{size, default is 15}

\item{family}{font family, default is Arial}

\item{bold}{TRUE}

\item{color}{black}

\item{italic}{FLASE}

\item{underlined}{FALSE}

\item{empty_line}{TRUE}

\item{vertical.align}{baseline}
}
\description{
Title
}
