% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add_image.R
\name{docx_add_image}
\alias{docx_add_image}
\title{add image}
\usage{
docx_add_image(
  image,
  height = 5,
  width = 5,
  style = "centered",
  title = NULL,
  foot = NULL
)
}
\arguments{
\item{image}{image name or ggplot2 object}

\item{height}{height, 5 is dafult}

\item{width}{width, 5 is default}

\item{style}{centered}

\item{title}{title for image}

\item{foot}{foot for image}
}
\description{
add image
}
